Consequents finished - model learning
The model is learning apropiately, sometimes the gradient explodes and un-learns. Main issue is that there is no way to regulate the learning rate on the LSE and RLSE params. Will proceed to design a regulatory system to reduce the lr as it approach the final result and to recuperate past best results.